requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
python-dotenv
tiktoken
langchain-community